CLOUDSTACK-8656: move cli test tool to separate file
